tests-ui/requirements.txt
Appium-Python-Client==1.0.2
assertpy==1.1
browserstack-local==1.2.2
configparser==5.0.1
cucumber-tag-expressions==3.0.0
diffimg==0.2.3
python-dotenv==0.15.0
Faker==4.14.2
gherkin-official==4.1.3
html5lib==1.1
imutils==0.5.3
opencv-python==4.4.0.46
pandas
py==1.9.0
pylint==2.6.0
pytest==6.1.2
pytest-base-url==1.4.2
pytest-bdd==4.0.1
pytest-rerunfailures==9.1.1
pytest-selenium==2.0.1
pytest-selenium-enhancer==1.6.1
pytest-testrail-client==1.0.8
pytest-xdist==2.1.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
scikit-image==0.17.2
selenium==3.141.0
Pillow==8.0.1
SSIM-PIL==1.0.10
webdrivermanager==0.9.0
